+++ tests/Stubs/TestRequests/WelcomeStoreRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Rakutentech\LaravelRequestDocs\Tests\TestRequests;
+namespace Rakutentech\LaravelRequestDocs\Tests\Stubs\TestRequests;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class WelcomeDeleteRequest extends FormRequest
+class WelcomeStoreRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -28,6 +28,7 @@ class WelcomeDeleteRequest extends FormRequest
     public function rules()
     {
         return [
+            'error'         => ['string', 'exists:' . $this->user->id],
             'message_param' => 'nullable|string',
         ];
     }
